+++ traffic_monitor/experimental/traffic_monitor/deliveryservice/stat.go
@@ -21,8 +21,6 @@ package deliveryservice
 
 import (
 	"fmt"
-	"net/url"
-	"strconv"
 	"time"
 
 	"github.com/apache/incubator-trafficcontrol/traffic_monitor/experimental/common/log"
@@ -32,42 +30,13 @@ import (
 	"github.com/apache/incubator-trafficcontrol/traffic_monitor/experimental/traffic_monitor/enum"
 	"github.com/apache/incubator-trafficcontrol/traffic_monitor/experimental/traffic_monitor/health"
 	"github.com/apache/incubator-trafficcontrol/traffic_monitor/experimental/traffic_monitor/peer"
-	"github.com/apache/incubator-trafficcontrol/traffic_monitor/experimental/traffic_monitor/srvhttp"
 	todata "github.com/apache/incubator-trafficcontrol/traffic_monitor/experimental/traffic_monitor/trafficopsdata"
 	to "github.com/apache/incubator-trafficcontrol/traffic_ops/client"
 )
 
 // TODO remove 'ds' and 'stat' from names
 
-// Stats is the JSON-serialisable representation of delivery service Stats. It maps delivery service names to individual stat objects.
-type Stats struct {
-	DeliveryService map[enum.DeliveryServiceName]dsdata.Stat `json:"deliveryService"`
-	Time            time.Time                                `json:"-"`
-}
-
-// Copy performs a deep copy of this Stats object.
-func (s Stats) Copy() Stats {
-	b := NewStats()
-	for k, v := range s.DeliveryService {
-		b.DeliveryService[k] = v.Copy()
-	}
-	b.Time = s.Time
-	return b
-}
-
-// Get returns the stats for the given delivery service, and whether it exists.
-func (s Stats) Get(name enum.DeliveryServiceName) (dsdata.StatReadonly, bool) {
-	ds, ok := s.DeliveryService[name]
-	return ds, ok
-}
-
-// NewStats creates a new Stats object, initializing any pointer members.
-// TODO rename to just 'New'?
-func NewStats() Stats {
-	return Stats{DeliveryService: map[enum.DeliveryServiceName]dsdata.Stat{}}
-}
-
-func setStaticData(dsStats Stats, dsServers map[enum.DeliveryServiceName][]enum.CacheName) Stats {
+func setStaticData(dsStats dsdata.Stats, dsServers map[enum.DeliveryServiceName][]enum.CacheName) dsdata.Stats {
 	for ds, stat := range dsStats.DeliveryService {
 		stat.CommonStats.CachesConfiguredNum.Value = int64(len(dsServers[ds]))
 		dsStats.DeliveryService[ds] = stat // TODO consider changing dsStats.DeliveryService[ds] to a pointer so this kind of thing isn't necessary; possibly more performant, as well
@@ -75,7 +44,7 @@ func setStaticData(dsStats Stats, dsServers map[enum.DeliveryServiceName][]enum.
 	return dsStats
 }
 
-func addAvailableData(dsStats Stats, crStates peer.Crstates, serverCachegroups map[enum.CacheName]enum.CacheGroupName, serverDs map[enum.CacheName][]enum.DeliveryServiceName, serverTypes map[enum.CacheName]enum.CacheType, precomputed map[enum.CacheName]cache.PrecomputedData) (Stats, error) {
+func addAvailableData(dsStats dsdata.Stats, crStates peer.Crstates, serverCachegroups map[enum.CacheName]enum.CacheGroupName, serverDs map[enum.CacheName][]enum.DeliveryServiceName, serverTypes map[enum.CacheName]enum.CacheType, precomputed map[enum.CacheName]cache.PrecomputedData) (dsdata.Stats, error) {
 	for cache, available := range crStates.Caches {
 		cacheGroup, ok := serverCachegroups[cache]
 		if !ok {
@@ -152,105 +121,18 @@ func addAvailableData(dsStats Stats, crStates peer.Crstates, serverCachegroups m
 	return dsStats, nil
 }
 
-// LastStats includes the previously recieved stats for DeliveryServices and Caches, the stat itself, when it was received, and the stat value per second.
-type LastStats struct {
-	DeliveryServices map[enum.DeliveryServiceName]LastDSStat
-	Caches           map[enum.CacheName]LastStatsData
-}
-
-// NewLastStats returns a new LastStats object, initializing internal pointer values.
-func NewLastStats() LastStats {
-	return LastStats{DeliveryServices: map[enum.DeliveryServiceName]LastDSStat{}, Caches: map[enum.CacheName]LastStatsData{}}
-}
-
-// Copy performs a deep copy of this LastStats object.
-func (a LastStats) Copy() LastStats {
-	b := NewLastStats()
-	for k, v := range a.DeliveryServices {
-		b.DeliveryServices[k] = v.Copy()
-	}
-	for k, v := range a.Caches {
-		b.Caches[k] = v
-	}
-	return b
-}
-
-// LastDSStat maps and aggregates the last stats received for the given delivery service to caches, cache groups, types, and total.
-// TODO figure a way to associate this type with StatHTTP, with which its members correspond.
-type LastDSStat struct {
-	Caches      map[enum.CacheName]LastStatsData
-	CacheGroups map[enum.CacheGroupName]LastStatsData
-	Type        map[enum.CacheType]LastStatsData
-	Total       LastStatsData
-}
-
-// Copy performs a deep copy of this LastDSStat object.
-func (a LastDSStat) Copy() LastDSStat {
-	b := LastDSStat{
-		CacheGroups: map[enum.CacheGroupName]LastStatsData{},
-		Type:        map[enum.CacheType]LastStatsData{},
-		Caches:      map[enum.CacheName]LastStatsData{},
-		Total:       a.Total,
-	}
-	for k, v := range a.CacheGroups {
-		b.CacheGroups[k] = v
-	}
-	for k, v := range a.Type {
-		b.Type[k] = v
-	}
-	for k, v := range a.Caches {
-		b.Caches[k] = v
-	}
-	return b
-}
-
-func newLastDSStat() LastDSStat {
-	return LastDSStat{
-		CacheGroups: map[enum.CacheGroupName]LastStatsData{},
-		Type:        map[enum.CacheType]LastStatsData{},
-		Caches:      map[enum.CacheName]LastStatsData{},
-	}
-}
-
-// LastStatsData contains the last stats and per-second calculations for bytes and status codes received from a cache.
-type LastStatsData struct {
-	Bytes     LastStatData
-	Status2xx LastStatData
-	Status3xx LastStatData
-	Status4xx LastStatData
-	Status5xx LastStatData
-}
-
-// Sum returns the Sum() of each member data with the given LastStatsData corresponding members
-func (a LastStatsData) Sum(b LastStatsData) LastStatsData {
-	return LastStatsData{
-		Bytes:     a.Bytes.Sum(b.Bytes),
-		Status2xx: a.Status2xx.Sum(b.Status2xx),
-		Status3xx: a.Status3xx.Sum(b.Status3xx),
-		Status4xx: a.Status4xx.Sum(b.Status4xx),
-		Status5xx: a.Status5xx.Sum(b.Status5xx),
-	}
-}
-
-// LastStatData contains the value, time it was received, and per-second calculation since the previous stat, for a stat from a cache.
-type LastStatData struct {
-	PerSec float64
-	Stat   int64
-	Time   time.Time
-}
-
-// Sum adds the PerSec and Stat of the given data to this object. Time is meaningless for the summed object, and is thus set to 0.
-func (a LastStatData) Sum(b LastStatData) LastStatData {
-	return LastStatData{
-		PerSec: a.PerSec + b.PerSec,
-		Stat:   a.Stat + b.Stat,
+func newLastDSStat() dsdata.LastDSStat {
+	return dsdata.LastDSStat{
+		CacheGroups: map[enum.CacheGroupName]dsdata.LastStatsData{},
+		Type:        map[enum.CacheType]dsdata.LastStatsData{},
+		Caches:      map[enum.CacheName]dsdata.LastStatsData{},
 	}
 }
 
 // BytesPerKilobit is the number of bytes in a kilobit.
 const BytesPerKilobit = 125
 
-func addLastStat(lastData LastStatData, newStat int64, newStatTime time.Time) (LastStatData, error) {
+func addLastStat(lastData dsdata.LastStatData, newStat int64, newStatTime time.Time) (dsdata.LastStatData, error) {
 	if newStat == lastData.Stat {
 		return lastData, nil
 	}
@@ -272,7 +154,7 @@ func addLastStat(lastData LastStatData, newStat int64, newStatTime time.Time) (L
 	return lastData, nil
 }
 
-func addLastStats(lastData LastStatsData, newStats dsdata.StatCacheStats, newStatsTime time.Time) (LastStatsData, error) {
+func addLastStats(lastData dsdata.LastStatsData, newStats dsdata.StatCacheStats, newStatsTime time.Time) (dsdata.LastStatsData, error) {
 	errs := []error{nil, nil, nil, nil, nil}
 	lastData.Bytes, errs[0] = addLastStat(lastData.Bytes, newStats.OutBytes.Value, newStatsTime)
 	lastData.Status2xx, errs[1] = addLastStat(lastData.Status2xx, newStats.Status2xx.Value, newStatsTime)
@@ -282,7 +164,7 @@ func addLastStats(lastData LastStatsData, newStats dsdata.StatCacheStats, newSta
 	return lastData, util.JoinErrors(errs)
 }
 
-func addLastStatsToStatCacheStats(s dsdata.StatCacheStats, l LastStatsData) dsdata.StatCacheStats {
+func addLastStatsToStatCacheStats(s dsdata.StatCacheStats, l dsdata.LastStatsData) dsdata.StatCacheStats {
 	s.Kbps.Value = l.Bytes.PerSec / BytesPerKilobit
 	s.Tps2xx.Value = l.Status2xx.PerSec
 	s.Tps3xx.Value = l.Status3xx.PerSec
@@ -293,10 +175,10 @@ func addLastStatsToStatCacheStats(s dsdata.StatCacheStats, l LastStatsData) dsda
 }
 
 // addLastDSStatTotals takes a LastDSStat with only raw `Caches` data, and calculates and sets the `CacheGroups`, `Type`, and `Total` data, and returns the augmented structure.
-func addLastDSStatTotals(lastStat LastDSStat, cachesReporting map[enum.CacheName]bool, serverCachegroups map[enum.CacheName]enum.CacheGroupName, serverTypes map[enum.CacheName]enum.CacheType) LastDSStat {
-	cacheGroups := map[enum.CacheGroupName]LastStatsData{}
-	cacheTypes := map[enum.CacheType]LastStatsData{}
-	total := LastStatsData{}
+func addLastDSStatTotals(lastStat dsdata.LastDSStat, cachesReporting map[enum.CacheName]bool, serverCachegroups map[enum.CacheName]enum.CacheGroupName, serverTypes map[enum.CacheName]enum.CacheType) dsdata.LastDSStat {
+	cacheGroups := map[enum.CacheGroupName]dsdata.LastStatsData{}
+	cacheTypes := map[enum.CacheType]dsdata.LastStatsData{}
+	total := dsdata.LastStatsData{}
 	for cacheName, cacheStats := range lastStat.Caches {
 		if !cachesReporting[cacheName] {
 			continue
@@ -322,7 +204,7 @@ func addLastDSStatTotals(lastStat LastDSStat, cachesReporting map[enum.CacheName
 }
 
 // addDSPerSecStats calculates and adds the per-second delivery service stats to both the Stats and LastStats structures, and returns the augmented structures.
-func addDSPerSecStats(dsName enum.DeliveryServiceName, stat dsdata.Stat, lastStats LastStats, dsStats Stats, dsStatsTime time.Time, serverCachegroups map[enum.CacheName]enum.CacheGroupName, serverTypes map[enum.CacheName]enum.CacheType, mc to.TrafficMonitorConfigMap, events []health.Event) (Stats, LastStats, []health.Event) {
+func addDSPerSecStats(dsName enum.DeliveryServiceName, stat dsdata.Stat, lastStats dsdata.LastStats, dsStats dsdata.Stats, dsStatsTime time.Time, serverCachegroups map[enum.CacheName]enum.CacheGroupName, serverTypes map[enum.CacheName]enum.CacheType, mc to.TrafficMonitorConfigMap, events health.ThreadsafeEvents) (dsdata.Stats, dsdata.LastStats) {
 	err := error(nil)
 	lastStat, lastStatExists := lastStats.DeliveryServices[dsName]
 	if !lastStatExists {
@@ -354,11 +236,11 @@ func addDSPerSecStats(dsName enum.DeliveryServiceName, stat dsdata.Stat, lastSta
 		stat.CommonStats.IsAvailable.Value = false
 		stat.CommonStats.IsHealthy.Value = false
 		stat.CommonStats.ErrorStr.Value = errStr
-		events = append(events, health.Event{Time: time.Now(), Unix: time.Now().Unix(), Description: errStr, Name: dsName.String(), Hostname: dsName.String(), Type: "Delivery Service", Available: stat.CommonStats.IsAvailable.Value})
+		events.Add(health.Event{Time: time.Now(), Unix: time.Now().Unix(), Description: errStr, Name: dsName.String(), Hostname: dsName.String(), Type: "Delivery Service", Available: stat.CommonStats.IsAvailable.Value})
 	}
 
 	dsStats.DeliveryService[dsName] = stat
-	return dsStats, lastStats, events
+	return dsStats, lastStats
 }
 
 // latestBytes returns the most recent OutBytes from the given cache results, and the time of that result. It assumes zero results are not valid, but nonzero results with errors are valid.
@@ -370,7 +252,7 @@ func latestBytes(p cache.PrecomputedData) (int64, time.Time, error) {
 }
 
 // addCachePerSecStats calculates the cache per-second stats, adds them to LastStats, and returns the augmented object.
-func addCachePerSecStats(cacheName enum.CacheName, precomputed cache.PrecomputedData, lastStats LastStats) LastStats {
+func addCachePerSecStats(cacheName enum.CacheName, precomputed cache.PrecomputedData, lastStats dsdata.LastStats) dsdata.LastStats {
 	outBytes, outBytesTime, err := latestBytes(precomputed) // it's ok if `latestBytes` returns 0s with an error, `addLastStat` will refrain from setting it (unless the previous calculation was nonzero, in which case it will error appropriately).
 	if err != nil {
 		log.Warnf("while computing delivery service data for cache %v: %v\n", cacheName, err)
@@ -393,22 +275,21 @@ func addCachePerSecStats(cacheName enum.CacheName, precomputed cache.Precomputed
 // we set the (new - old) / lastChangedTime as the KBPS, and update the recorded LastChangedTime and LastChangedValue
 //
 // TODO handle ATS byte rolling (when the `out_bytes` overflows back to 0)
-func addPerSecStats(precomputed map[enum.CacheName]cache.PrecomputedData, dsStats Stats, lastStats LastStats, dsStatsTime time.Time, serverCachegroups map[enum.CacheName]enum.CacheGroupName, serverTypes map[enum.CacheName]enum.CacheType, mc to.TrafficMonitorConfigMap, events []health.Event) (Stats, LastStats, []health.Event) {
+func addPerSecStats(precomputed map[enum.CacheName]cache.PrecomputedData, dsStats dsdata.Stats, lastStats dsdata.LastStats, dsStatsTime time.Time, serverCachegroups map[enum.CacheName]enum.CacheGroupName, serverTypes map[enum.CacheName]enum.CacheType, mc to.TrafficMonitorConfigMap, events health.ThreadsafeEvents) (dsdata.Stats, dsdata.LastStats) {
 	for dsName, stat := range dsStats.DeliveryService {
-		dsStats, lastStats, events = addDSPerSecStats(dsName, stat, lastStats, dsStats, dsStatsTime, serverCachegroups, serverTypes, mc, events)
+		dsStats, lastStats = addDSPerSecStats(dsName, stat, lastStats, dsStats, dsStatsTime, serverCachegroups, serverTypes, mc, events)
 	}
 	for cacheName, precomputedData := range precomputed {
 		lastStats = addCachePerSecStats(cacheName, precomputedData, lastStats)
 	}
 
-	return dsStats, lastStats, events
+	return dsStats, lastStats
 }
 
 // CreateStats aggregates and creates statistics from given precomputed stat history. It returns the created stats, information about these stats necessary for the next calculation, and any error.
-func CreateStats(precomputed map[enum.CacheName]cache.PrecomputedData, toData todata.TOData, crStates peer.Crstates, lastStats LastStats, now time.Time, mc to.TrafficMonitorConfigMap) (Stats, LastStats, []health.Event, error) {
-	var events []health.Event
+func CreateStats(precomputed map[enum.CacheName]cache.PrecomputedData, toData todata.TOData, crStates peer.Crstates, lastStats dsdata.LastStats, now time.Time, mc to.TrafficMonitorConfigMap, events health.ThreadsafeEvents) (dsdata.Stats, dsdata.LastStats, error) {
 	start := time.Now()
-	dsStats := NewStats()
+	dsStats := dsdata.NewStats()
 	for deliveryService := range toData.DeliveryServiceServers {
 		if deliveryService == "" {
 			log.Errorf("EMPTY CreateStats deliveryService")
@@ -420,7 +301,7 @@ func CreateStats(precomputed map[enum.CacheName]cache.PrecomputedData, toData to
 	var err error
 	dsStats, err = addAvailableData(dsStats, crStates, toData.ServerCachegroups, toData.ServerDeliveryServices, toData.ServerTypes, precomputed) // TODO move after stat summarisation
 	if err != nil {
-		return dsStats, lastStats, events, fmt.Errorf("Error getting Cache availability data: %v", err)
+		return dsStats, lastStats, fmt.Errorf("Error getting Cache availability data: %v", err)
 	}
 
 	for server, precomputedData := range precomputed {
@@ -457,82 +338,10 @@ func CreateStats(precomputed map[enum.CacheName]cache.PrecomputedData, toData to
 		}
 	}
 
-	perSecStats, lastStats, events := addPerSecStats(precomputed, dsStats, lastStats, now, toData.ServerCachegroups, toData.ServerTypes, mc, events)
+	perSecStats, lastStats := addPerSecStats(precomputed, dsStats, lastStats, now, toData.ServerCachegroups, toData.ServerTypes, mc, events)
 	log.Infof("CreateStats took %v\n", time.Since(start))
 	perSecStats.Time = time.Now()
-	return perSecStats, lastStats, events, nil
-}
-
-func addStatCacheStats(s *dsdata.StatsOld, c dsdata.StatCacheStats, deliveryService enum.DeliveryServiceName, prefix string, t int64, filter dsdata.Filter) *dsdata.StatsOld {
-	add := func(name, val string) {
-		if filter.UseStat(name) {
-			// This is for compatibility with the Traffic Monitor 1.0 API.
-			// TODO abstract this? Or deprecate and remove it?
-			if name == "isAvailable" || name == "error-string" {
-				s.DeliveryService[deliveryService][dsdata.StatName("location."+prefix+name)] = []dsdata.StatOld{dsdata.StatOld{Time: t, Value: val}}
-			} else {
-				s.DeliveryService[deliveryService][dsdata.StatName(prefix+name)] = []dsdata.StatOld{dsdata.StatOld{Time: t, Value: val}}
-			}
-		}
-	}
-	add("out_bytes", strconv.Itoa(int(c.OutBytes.Value)))
-	add("isAvailable", fmt.Sprintf("%t", c.IsAvailable.Value))
-	add("status_5xx", strconv.Itoa(int(c.Status5xx.Value)))
-	add("status_4xx", strconv.Itoa(int(c.Status4xx.Value)))
-	add("status_3xx", strconv.Itoa(int(c.Status3xx.Value)))
-	add("status_2xx", strconv.Itoa(int(c.Status2xx.Value)))
-	add("in_bytes", strconv.Itoa(int(c.InBytes.Value)))
-	add("kbps", strconv.Itoa(int(c.Kbps.Value)))
-	add("tps_5xx", fmt.Sprintf("%f", c.Tps5xx.Value))
-	add("tps_4xx", fmt.Sprintf("%f", c.Tps4xx.Value))
-	add("tps_3xx", fmt.Sprintf("%f", c.Tps3xx.Value))
-	add("tps_2xx", fmt.Sprintf("%f", c.Tps2xx.Value))
-	add("error-string", c.ErrorString.Value)
-	add("tps_total", fmt.Sprintf("%f", c.TpsTotal.Value))
-	return s
-}
-
-func addCommonData(s *dsdata.StatsOld, c *dsdata.StatCommon, deliveryService enum.DeliveryServiceName, t int64, filter dsdata.Filter) *dsdata.StatsOld {
-	add := func(name string, val interface{}) {
-		if filter.UseStat(name) {
-			s.DeliveryService[deliveryService][dsdata.StatName(name)] = []dsdata.StatOld{dsdata.StatOld{Time: t, Value: val}}
-		}
-	}
-	add("caches-configured", fmt.Sprintf("%d", c.CachesConfiguredNum.Value))
-	add("caches-reporting", fmt.Sprintf("%d", len(c.CachesReporting)))
-	add("error-string", c.ErrorStr.Value)
-	add("status", c.StatusStr.Value)
-	add("isHealthy", fmt.Sprintf("%t", c.IsHealthy.Value))
-	add("isAvailable", fmt.Sprintf("%t", c.IsAvailable.Value))
-	add("caches-available", fmt.Sprintf("%d", c.CachesAvailableNum.Value))
-	add("disabledLocations", c.CachesDisabled)
-	return s
-}
-
-// JSON returns an object formatted as expected to be serialized to JSON and served.
-func (s Stats) JSON(filter dsdata.Filter, params url.Values) dsdata.StatsOld {
-	// TODO fix to be the time calculated, not the time requested
-	now := s.Time.UnixNano() / int64(time.Millisecond) // Traffic Monitor 1.0 API is 'ms since the epoch'
-	jsonObj := &dsdata.StatsOld{
-		CommonAPIData:   srvhttp.GetCommonAPIData(params, time.Now()),
-		DeliveryService: map[enum.DeliveryServiceName]map[dsdata.StatName][]dsdata.StatOld{},
-	}
-
-	for deliveryService, stat := range s.DeliveryService {
-		if !filter.UseDeliveryService(deliveryService) {
-			continue
-		}
-		jsonObj.DeliveryService[deliveryService] = map[dsdata.StatName][]dsdata.StatOld{}
-		jsonObj = addCommonData(jsonObj, &stat.CommonStats, deliveryService, now, filter)
-		for cacheGroup, cacheGroupStats := range stat.CacheGroups {
-			jsonObj = addStatCacheStats(jsonObj, cacheGroupStats, deliveryService, "location."+string(cacheGroup)+".", now, filter)
-		}
-		for cacheType, typeStats := range stat.Types {
-			jsonObj = addStatCacheStats(jsonObj, typeStats, deliveryService, "type."+cacheType.String()+".", now, filter)
-		}
-		jsonObj = addStatCacheStats(jsonObj, stat.TotalStats, deliveryService, "total.", now, filter)
-	}
-	return *jsonObj
+	return perSecStats, lastStats, nil
 }
 
 func getDsErrString(dsName enum.DeliveryServiceName, dsStats dsdata.StatCacheStats, monitorConfig to.TrafficMonitorConfigMap) string {
