+++ traffic_monitor/experimental/traffic_monitor/deliveryservicedata/stat.go
@@ -21,7 +21,9 @@ package deliveryservicedata // TODO rename?
 
 import (
 	"errors"
+	"fmt"
 	"net/url"
+	"strconv"
 	"time"
 
 	"github.com/apache/incubator-trafficcontrol/traffic_monitor/experimental/traffic_monitor/enum"
@@ -288,4 +290,199 @@ func (a Stat) Type(name enum.CacheType) (StatCacheStats, bool) {
 // Total returns the aggregated total data in this stat. It is part of the StatCommonReadonly interface.
 func (a Stat) Total() StatCacheStats {
 	return a.TotalStats
+}
+
+// Stats is the JSON-serialisable representation of delivery service Stats. It maps delivery service names to individual stat objects.
+type Stats struct {
+	DeliveryService map[enum.DeliveryServiceName]Stat `json:"deliveryService"`
+	Time            time.Time                         `json:"-"`
+}
+
+// Copy performs a deep copy of this Stats object.
+func (s Stats) Copy() Stats {
+	b := NewStats()
+	for k, v := range s.DeliveryService {
+		b.DeliveryService[k] = v.Copy()
+	}
+	b.Time = s.Time
+	return b
+}
+
+// Get returns the stats for the given delivery service, and whether it exists.
+func (s Stats) Get(name enum.DeliveryServiceName) (StatReadonly, bool) {
+	ds, ok := s.DeliveryService[name]
+	return ds, ok
+}
+
+// JSON returns an object formatted as expected to be serialized to JSON and served.
+func (s Stats) JSON(filter Filter, params url.Values) StatsOld {
+	// TODO fix to be the time calculated, not the time requested
+	now := s.Time.UnixNano() / int64(time.Millisecond) // Traffic Monitor 1.0 API is 'ms since the epoch'
+	jsonObj := &StatsOld{
+		CommonAPIData:   srvhttp.GetCommonAPIData(params, time.Now()),
+		DeliveryService: map[enum.DeliveryServiceName]map[StatName][]StatOld{},
+	}
+
+	for deliveryService, stat := range s.DeliveryService {
+		if !filter.UseDeliveryService(deliveryService) {
+			continue
+		}
+		jsonObj.DeliveryService[deliveryService] = map[StatName][]StatOld{}
+		jsonObj = addCommonData(jsonObj, &stat.CommonStats, deliveryService, now, filter)
+		for cacheGroup, cacheGroupStats := range stat.CacheGroups {
+			jsonObj = addStatCacheStats(jsonObj, cacheGroupStats, deliveryService, "location."+string(cacheGroup)+".", now, filter)
+		}
+		for cacheType, typeStats := range stat.Types {
+			jsonObj = addStatCacheStats(jsonObj, typeStats, deliveryService, "type."+cacheType.String()+".", now, filter)
+		}
+		jsonObj = addStatCacheStats(jsonObj, stat.TotalStats, deliveryService, "total.", now, filter)
+	}
+	return *jsonObj
+}
+
+// NewStats creates a new Stats object, initializing any pointer members.
+// TODO rename to just 'New'?
+func NewStats() Stats {
+	return Stats{DeliveryService: map[enum.DeliveryServiceName]Stat{}}
+}
+
+// LastStats includes the previously recieved stats for DeliveryServices and Caches, the stat itself, when it was received, and the stat value per second.
+type LastStats struct {
+	DeliveryServices map[enum.DeliveryServiceName]LastDSStat
+	Caches           map[enum.CacheName]LastStatsData
+}
+
+// NewLastStats returns a new LastStats object, initializing internal pointer values.
+func NewLastStats() LastStats {
+	return LastStats{DeliveryServices: map[enum.DeliveryServiceName]LastDSStat{}, Caches: map[enum.CacheName]LastStatsData{}}
+}
+
+// Copy performs a deep copy of this LastStats object.
+func (a LastStats) Copy() LastStats {
+	b := NewLastStats()
+	for k, v := range a.DeliveryServices {
+		b.DeliveryServices[k] = v.Copy()
+	}
+	for k, v := range a.Caches {
+		b.Caches[k] = v
+	}
+	return b
+}
+
+// LastDSStat maps and aggregates the last stats received for the given delivery service to caches, cache groups, types, and total.
+// TODO figure a way to associate this type with StatHTTP, with which its members correspond.
+type LastDSStat struct {
+	Caches      map[enum.CacheName]LastStatsData
+	CacheGroups map[enum.CacheGroupName]LastStatsData
+	Type        map[enum.CacheType]LastStatsData
+	Total       LastStatsData
+}
+
+// Copy performs a deep copy of this LastDSStat object.
+func (a LastDSStat) Copy() LastDSStat {
+	b := LastDSStat{
+		CacheGroups: map[enum.CacheGroupName]LastStatsData{},
+		Type:        map[enum.CacheType]LastStatsData{},
+		Caches:      map[enum.CacheName]LastStatsData{},
+		Total:       a.Total,
+	}
+	for k, v := range a.CacheGroups {
+		b.CacheGroups[k] = v
+	}
+	for k, v := range a.Type {
+		b.Type[k] = v
+	}
+	for k, v := range a.Caches {
+		b.Caches[k] = v
+	}
+	return b
+}
+
+func newLastDSStat() LastDSStat {
+	return LastDSStat{
+		CacheGroups: map[enum.CacheGroupName]LastStatsData{},
+		Type:        map[enum.CacheType]LastStatsData{},
+		Caches:      map[enum.CacheName]LastStatsData{},
+	}
+}
+
+// LastStatsData contains the last stats and per-second calculations for bytes and status codes received from a cache.
+type LastStatsData struct {
+	Bytes     LastStatData
+	Status2xx LastStatData
+	Status3xx LastStatData
+	Status4xx LastStatData
+	Status5xx LastStatData
+}
+
+// Sum returns the Sum() of each member data with the given LastStatsData corresponding members
+func (a LastStatsData) Sum(b LastStatsData) LastStatsData {
+	return LastStatsData{
+		Bytes:     a.Bytes.Sum(b.Bytes),
+		Status2xx: a.Status2xx.Sum(b.Status2xx),
+		Status3xx: a.Status3xx.Sum(b.Status3xx),
+		Status4xx: a.Status4xx.Sum(b.Status4xx),
+		Status5xx: a.Status5xx.Sum(b.Status5xx),
+	}
+}
+
+// LastStatData contains the value, time it was received, and per-second calculation since the previous stat, for a stat from a cache.
+type LastStatData struct {
+	PerSec float64
+	Stat   int64
+	Time   time.Time
+}
+
+// Sum adds the PerSec and Stat of the given data to this object. Time is meaningless for the summed object, and is thus set to 0.
+func (a LastStatData) Sum(b LastStatData) LastStatData {
+	return LastStatData{
+		PerSec: a.PerSec + b.PerSec,
+		Stat:   a.Stat + b.Stat,
+	}
+}
+
+func addCommonData(s *StatsOld, c *StatCommon, deliveryService enum.DeliveryServiceName, t int64, filter Filter) *StatsOld {
+	add := func(name string, val interface{}) {
+		if filter.UseStat(name) {
+			s.DeliveryService[deliveryService][StatName(name)] = []StatOld{StatOld{Time: t, Value: val}}
+		}
+	}
+	add("caches-configured", fmt.Sprintf("%d", c.CachesConfiguredNum.Value))
+	add("caches-reporting", fmt.Sprintf("%d", len(c.CachesReporting)))
+	add("error-string", c.ErrorStr.Value)
+	add("status", c.StatusStr.Value)
+	add("isHealthy", fmt.Sprintf("%t", c.IsHealthy.Value))
+	add("isAvailable", fmt.Sprintf("%t", c.IsAvailable.Value))
+	add("caches-available", fmt.Sprintf("%d", c.CachesAvailableNum.Value))
+	add("disabledLocations", c.CachesDisabled)
+	return s
+}
+
+func addStatCacheStats(s *StatsOld, c StatCacheStats, deliveryService enum.DeliveryServiceName, prefix string, t int64, filter Filter) *StatsOld {
+	add := func(name, val string) {
+		if filter.UseStat(name) {
+			// This is for compatibility with the Traffic Monitor 1.0 API.
+			// TODO abstract this? Or deprecate and remove it?
+			if name == "isAvailable" || name == "error-string" {
+				s.DeliveryService[deliveryService][StatName("location."+prefix+name)] = []StatOld{StatOld{Time: t, Value: val}}
+			} else {
+				s.DeliveryService[deliveryService][StatName(prefix+name)] = []StatOld{StatOld{Time: t, Value: val}}
+			}
+		}
+	}
+	add("out_bytes", strconv.Itoa(int(c.OutBytes.Value)))
+	add("isAvailable", fmt.Sprintf("%t", c.IsAvailable.Value))
+	add("status_5xx", strconv.Itoa(int(c.Status5xx.Value)))
+	add("status_4xx", strconv.Itoa(int(c.Status4xx.Value)))
+	add("status_3xx", strconv.Itoa(int(c.Status3xx.Value)))
+	add("status_2xx", strconv.Itoa(int(c.Status2xx.Value)))
+	add("in_bytes", strconv.Itoa(int(c.InBytes.Value)))
+	add("kbps", strconv.Itoa(int(c.Kbps.Value)))
+	add("tps_5xx", fmt.Sprintf("%f", c.Tps5xx.Value))
+	add("tps_4xx", fmt.Sprintf("%f", c.Tps4xx.Value))
+	add("tps_3xx", fmt.Sprintf("%f", c.Tps3xx.Value))
+	add("tps_2xx", fmt.Sprintf("%f", c.Tps2xx.Value))
+	add("error-string", c.ErrorString.Value)
+	add("tps_total", fmt.Sprintf("%f", c.TpsTotal.Value))
+	return s
 }
